requirements.txt
Pillow
fastapi
uvicorn
python-multipart
httpx
tensorflow
pandas